Move test from DFS verifier to BProgramJsProxy Add test for eventset and not just BEvents.
