+++ src/test/java/il/ac/bgu/cs/bp/bpjs/analysis/DfsBProgramVerifierTest.java
@@ -105,16 +105,7 @@ public class DfsBProgramVerifierTest {
         
     }
 
-    @Test
-    public void DeadlockSameThread() throws Exception{
-        BProgram bpr = new SingleResourceBProgram("bpsync-blockrequest.js");
-        DfsBProgramVerifier sut = new DfsBProgramVerifier();
-        sut.setVisitedNodeStore(new ForgetfulVisitedStateStore());
-        VerificationResult res = sut.verify(bpr);
-        assertTrue(res.isCounterExampleFound());
-        assertEquals(VerificationResult.ViolationType.Deadlock, res.getViolationType());
-        assertEquals("sampleEvent", traceEventNamesString(res.getCounterExampleTrace(), ""));
-    }
+
     
     @Test
     public void testTwoSimpleBThreads() throws Exception {
